rework default scheduler settings
